fix(build/tests): updates build to emit to root directory
The reasoning behind this is the "type": "module" declaration only works if the files are in the
same directory as the package.json with this declaration. This commit also disables testing as ESM
is not supported in Jest yet and I threw the kitchen sink at the problem but could not get it
working. I will open an issue and subscribe to this issue on the Jest repository.